update lda code
